APP_WEBPAGE/requirements.txt
gspread==6.1.0
oauth2client==4.1.3
pandas==1.4.4
st_annotated_text==4.0.1
streamlit==1.32.2
streamlit_authenticator==0.1.5
streamlit_card==1.0.0
